backend/requirements-dev.txt
pytest
pytest-asyncio
httpx
fastapi-cli 